Create OConnor_Contribution.cpp
I thought that it would simplify your code more by combining the "insert first item" and "insert item function into just one function that inserts an item. I wrote the code below that just uses a simple if statement that checks if the data element is on the first element or not.